working cleanup() messy
