+++ src/bot/bot.py
@@ -19,7 +19,7 @@ import discord
 from src.bot import initialize, message_handler, media_handler
 from src.core import constants
 from src.memory import memory
-from src.voice import voice
+from src.voice.voice import Voice
 
 # Configuring intents (this is crucial!!)
 intents = discord.Intents.default() # default intents
@@ -30,10 +30,15 @@ bot = discord.Bot(intents=intents)
 
 def start():
     """
-    Starts the bot.
+    Adds any cogs and starts the bot.
     """
     try:
+        # Add cogs
+        bot.add_cog(Voice(bot)) # Voice functionality
+
+        # Start bot proper
         bot.run(constants.DISCORD)
+
     except:
         print('Unable to initialize bot.')
 
@@ -45,6 +50,7 @@ async def on_ready():
     # Add Bot user ID to constants
     constants.BOT_ID = bot.user.id
 
+    # Set discord presence
     await initialize.set_presence(bot=bot)
 
     # Print ASCII in terminal to signify ready
@@ -74,17 +80,6 @@ async def on_message(message):
         else:
             await message_handler.reply(message=message)
 
-# @bot.event
-# async def on_voice_state_update(member, before, after):
-#     """
-#     Disconnect handler.
-#     """
-#     # Check if the bot is the member and has left a voice channel
-#     if member == bot.user and after.channel is None:
-#         guild_id = member.guild.id  # Access guild ID through the member object
-#         # Check if the guild is in the voice connections and remove it if present
-#         voice.connections.pop(guild_id, None)
-
 '''
 /Slash Commands
 '''
@@ -95,92 +90,15 @@ async def ping(ctx):
 #######################################################################
 
 '''
-REWRITE LOGIC:
+THINGS TO WORK ON IN THIS REFACTORING:
 
-1. One VC command.
-2. All other features need VC command to be active.
+1. One primary vc() command.
+2. All other features like read() need VC command to be active, and add onto vc()
 3. Rethink alone() function - it is TOO chonky
 4. Too much verbosity - rethink upon python standards.
 5. Don't worry about error handling to begin with - first have a stable implementation.
 '''
 
-@bot.slash_command(description='Read text aloud.')
-async def read(ctx, text: str):
-    # Ensure it's in a guild
-    if ctx.guild is None:
-        await ctx.respond(f"i can only vc in servers", ephemeral=True)
-        return
 
-    # Create/retrieve voice client
-    voice_client = await voice.connect(ctx=ctx)
 
-    if not voice_client or not voice_client.is_connected():
-        await ctx.respond(f"i think it broke", ephemeral=True)
-        voice.connections.pop(ctx.guild.id, None)
-        return
-    
-    # Respond ephemerally with text to be read in quotes.
-    await ctx.respond(f'> “{text}”', ephemeral=True)
 
-    # Begin read()
-    await voice.read(ctx=ctx, voice_client=voice_client, text=text)
-
-    # Handover to alone() if required
-    if (
-        not voice.connections.get(ctx.guild.id, {}).get('state') and
-        not voice.connections.get(ctx.guild.id, {}).get('alone_loop')
-        ):
-        await voice.alone(ctx=ctx, voice_client=voice_client)
-
-@bot.slash_command(description='Start voice call.')
-async def vc(ctx):
-    # Ensure it's in a guild
-    if ctx.guild is None:
-        await ctx.respond(f"i can only vc in servers", ephemeral=True)
-        return
-
-    # Create/retrieve voice client
-    voice_client = await voice.connect(ctx=ctx)
-
-    if not voice_client or not voice_client.is_connected():
-        await ctx.respond(f"i think it broke", ephemeral=True)
-        voice.connections.pop(ctx.guild.id, None)
-        return 
-
-    # Check if the bot is already in VC due to previous /vc command
-    if voice.connections.get(ctx.guild.id, {}).get('state'):
-        await ctx.respond("im already in vc", ephemeral=True)
-
-    elif not voice.connections.get(ctx.guild.id, {}).get('alone_loop'):
-        # Update the state to indicate the bot is in the VC due to /vc command
-        voice.connections[ctx.guild.id]['state'] = True
-
-        # Declare that the request has been acknowledged
-        await ctx.respond("ok ready now", ephemeral=True)
-
-        # Hand-off to alone() loop for idle disconnect
-        await voice.alone(ctx=ctx, voice_client=voice_client)
-
-@bot.slash_command(description='Reset the voice state in this server.')
-async def reset(ctx):
-    # Ensure it's in a guild
-    if ctx.guild is None:
-        await ctx.respond(f"i can only vc in servers", ephemeral=True)
-        return
-
-    guild_id = ctx.guild.id
-
-    # Check if the guild is in the voice connections and remove it if present
-    if guild_id in voice.connections:
-
-        # If the bot is connected to a voice channel, disconnect it first
-        voice_client = voice.connections[guild_id].get('voice_client')
-        if voice_client and voice_client.is_connected():
-            await voice_client.disconnect()
-
-        # Remove the guild from connections
-        voice.connections.pop(guild_id, None)
-
-        await ctx.respond("voice state has been reset")
-    else:
-        await ctx.respond("there isnt anything to reset")
